+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ﻿## CHANGELOG
 
+------------------------------------------------
+
+## Version 3.8.0
+   ###### Date: 20-Oct-2020
+   - **Asset**   includeFallback support added
+   - **AssetQuery**   includeFallback support added
+   - **Entry**   includeFallback support added
+   - **Query**   includeFallback support added
+
+------------------------------------------------
+
 ### Version: 3.7.0
 ###### Date: 15-Nov-2019
    - [Stack]: Added support for function getContentType()